requirements.txt
flask
PyPDF2
sentence-transformers
faiss-cpu
gunicorn
